pyproject.toml
[project]
name = "clinical-trials-scout"
version = "0.1.0"
description = "Clinical Trial Scout - FastAPI chatbot with Anthropic Claude and ClinicalTrials.gov API integration"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "jinja2>=3.1.6",
    "jinja2-fragments>=1.10.0",
    "python-multipart>=0.0.20",
    "tortoise-orm[asyncpg]>=0.25.1",
    "asyncpg>=0.30.0",
    "aerich>=0.8.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "litellm>=1.80.9",
    "httpx[http2]>=0.28.1",
    "markdown2>=2.5.4",
    "itsdangerous>=2.1.0",
]

[tool.aerich]
tortoise_orm = "app.config.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."